continued on fight structure
