+++ Assembly-CSharp.csproj
@@ -63,10 +63,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Assets\Scripts\CombatStuff\Actions\BasicAction.cs" />
     <Compile Include="Assets\Scripts\CombatStuff\Actions\IgnoreStatus.cs" />
     <Compile Include="Assets\Scripts\CombatStuff\Actions\StatChange.cs" />
     <Compile Include="Assets\Scripts\CombatStuff\Actions\StrengthMod.cs" />
+    <Compile Include="Assets\Scripts\CombatStuff\Actions\ThroughDef.cs" />
     <Compile Include="Assets\Scripts\CombatStuff\Fighter.cs" />
     <Compile Include="Assets\Scripts\CombatStuff\PlayerFighter.cs" />
     <Compile Include="Assets\Scripts\EnemyScript\EnemyMoveTypes\CardinalMove.cs" />
