--- Assets/Scripts/CombatStuff/Actions/BasicAction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 96500a9f1ceb59142bc4837f614a05f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
