+++ Assets/Scripts/CombatStuff/Actions/StatChange.cs
@@ -2,17 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class StatChange : MonoBehaviour
+public interface StatChange
 {
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
+    
 }
